pyproject.toml
[tool.poetry]
name = "deeplearning"
version = "0.1.0"
description = "ITBA Deep Learning Course"
authors = ["Leonardo Millan <leonardo.a.millan@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11 < 3.12"
pandas = "^2.2.2"
matplotlib = "^3.8.4"
seaborn = "^0.13.2"
geoplot = "^0.5.1"
pyzipcode = "^3.0.1"
numba = "^0.59.1"
mlflow = "^2.14.1"

# Issue between poetry and tensorflow metadata since >=2.11
# This is a temporary workaround
# related to https://github.com/python-poetry/poetry/issues/8271
# Inspired from https://github.com/tensorflow/tensorflow/blob/adb39b04e9cb116df4659a7e2de9eea27e62f25c/tensorflow/tools/pip_package/setup.py#L148-L162
tensorflow = {version = "^2.13.0" }
tensorflow-intel = { version = "^2.13.0", platform = "win32" }
tensorflow-io-gcs-filesystem = [
    { version = "< 0.32.0", markers = "platform_system == 'Windows'" }
]
scikit-learn = "^1.5.1"
keras = "^3.4.1"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
psycopg2 = "^2.9.9"
dependency-injector = "^4.41.0"
flask-restx = "^1.3.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
jupyterlab = "^4.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
